Add a Emacs .dir-locals.el file to locally use TABs in indentation
See
https://www.gnu.org/software/emacs/manual/html_node/emacs/Directory-Variables.html
for how this file works.

This is GNU Emacs specific. Other text editors would have to have
their own per-directory configuration to change from the defacto
standard indentation of 4 spaces and no TABs.

I did this just to comply with the original authors use of TABs in
indentation, although I disagree with it. If I needed to do more
indepth development on this project, I would want to reindent the code
to a defacto standard of 4 spaces per indentation level and no TABs. But
I'm doing this so that my debug code, that I am planning to add later,
aligns properly, just in case I need to engage the author further on
any issues I find.